requirements.txt
google-generativeai
streamlit
streamlit-mic-recorder
openai
firebase-admin
matplotlib
textblob